+++ example/bare/src/types.ts
@@ -4,6 +4,7 @@ export type AppStackParamList = {
   Standard: undefined;
   Test: undefined;
   ModalStack: undefined;
+  TestStack: undefined;
 };
 
 export type ModalStackParamList = {
@@ -15,6 +16,8 @@ export type SheetStackParamList = {
   SheetHomeStack: undefined;
   Details: undefined;
   Settings: undefined;
+  Profile: undefined;
+  Notifications: undefined;
 };
 
 export type SheetHomeStackParamList = {